refactor!: components folder structure
