+++ webapp/app/admin/trainee-manager.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import type { Profile } from "@backend/database/schema";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
-import { DataTable } from "~/components/ui/data-table";
+import { DataTable } from "~/components/data-table";
 import { PageHeader } from "~/components/page-header";
 import { TraineeReservations } from "./trainee-manager/reservations";
 import { useTRPC } from "~/utils/trpc";
